backend/requirements.txt
flask
flask-cors
opencv-python-headless
numpy
pillow
rembg
onnxruntime
mediapipe==0.10.9
protobuf==3.20.3
gunicorn